Used .setState to change the state.
Changed the name of people in the Persons component by passing an argument in the switchNameHandler function. Used the switchNameHandler function as a property and changed the argument using an arrow function and bind(). Bind() is the preferred way of doing this because an arrow function is less efficient. Also did some styling, both on a separate stylesheet that was imported and with inline styling.